requirements.txt
streamlit
pandas
numpy
matplotlib

langgraph
openai
python-dotenv
yfinance
python-dotenv
pypdf
faiss-cpu

